requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==4.2.11
djangorestframework==3.15.1
gunicorn==20.1.0
psycopg2-binary==2.9.10
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
